feat: drop sse support, simplify with stateless/streamable http
